Texts creation and link with DataBase done, modification to application.properties to applies (try to update body into varhar(255) -> incompatible with words limits condition) Theme.Id deleted in createText.html because bug (didn't manage to resolve that at this moment) singleText.html not modified for now -> next step with nobody did it pom.xml update with new dependencies for Hibernate (the query system of jakarta)
Took 4 hours 43 minutes

+++ src/java/Utils/TextRepository.java
@@ -10,13 +10,15 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TextRepository extends JpaRepository<Text, Long> {
+public interface TextRepository extends JpaRepository<Text, Integer> {
     //Optional<Text> findByText_title(String text_title);
     //Optional<Text> findByBody(String body);
     Optional<Text> findByStatus(String status);
+
     //Optional<Text> findByText_submit(Boolean text_submit);
     //Optional<Text> findBySubmittedAt(Timestamp submitted_at);
     Optional<Text> findByReported(Boolean reported);
+
     Optional<Text> findByDisqualified(Boolean disqualified);
 
     @Query("SELECT t FROM Text t JOIN FETCH t.user u WHERE t.challenge.challengeId = :challengeId AND t.textSubmit = true")
